Fix job support conversion break due to previous change
